requirements.txt
langchain
langchain_community
openai-whisper
faiss-cpu